charlie_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.0
python-dotenv==1.0.1
channels==4.0.0
daphne==4.0.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
psycopg-binary==3.2.5
whitenoise==6.6.0
gunicorn==21.2.0
django-extensions==3.2.3
